+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [markdownlint](https://dlaa.me/markdownlint/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added to Unreleased
+
+- Added `hello_world.py` that outputs "Hello, world!" and the current timestamp
+- Added IBM stock price display to `hello_world.py` output using yfinance
+
 ## [1.0.0] - yyyy-mm-dd
 
 ### Added to 1.0.0